20 new releases, take 2
I'm resubmitting this. I've actioned some of the issues you raised but not all of them:

- Technical errors like empty tags have hopefully been fixed, trailing newlines from pastes were the main culprit here. I checked the new file in notepad before running the json script.
- I have not made any changes to generic notes, they aren't perfect but can be removed in bulk once proper functionality is added to the pages to highlight dependencies/increased limit/bsp2.
- I changed the handwritten note for Immortal Lock to something a bit more generic. I would argue a gentle nudge towards the readme is courteous for this map in particular.
- I have tweaked a number of descriptions that had some possibly-subjective wording. Immortal Lock & Tomb of Thunder in particular.
- I have not improved the short, lazy descriptions for GraveYardKaiser's maps as imho they are entirely consistent with the amount of effort put into the maps. 'Unsealed map in a variety of textures' is literally all I can say about two of them. If you object to them then I would ask you to just edit them to 'No description added'.
- I removed one use of the word 'new' you objected to but there are others I didn't, e.g. referring to 'new monsters and weapons', I like to think people wouldn't interpret the word as temporal but as 'custom created for this pack'. I figure if I was telling someone about, say, Scourge of Armagon, I'd probably still describe it as having 'new monsters', despite them being currently 27 years old!

On descriptions in general, I would personally prefer them as despite the DB going in a more machine-readable direction, the website is still used a lot by humans. Maybe a disclaimer could be added saying descriptions have been added by voluntary contributors and are do not necessarily reflect the views of Quaddicted? And a note that amendments can be made upon request? I think a lot of the flak can probably be avoided by just appearing to be flexible and valuing feedback.

But yeah, in the long run creators will hopefully be submitting their own descriptions. So like the notes, this is probably just a temporary thing in the name of consistency. At some point in the future there probably needs to be a big exercise in taking excerpts from readmes to use as user descriptions for all the stuff currently there, and the official descriptions slowly retired. Kicking that down the line though :D

On tags, I think when I started doing these batch updates I thought 'descriptive tags can just be added by users later on' like they were on the old site. Hopefully the new one gets to that point eventually and this job can be farmed out a bit.