New version of grunt grenadin
Adding a new version of a 19-year old map at the author's request... very manual as I haven't even begun to streamline the updating process

Up until now I've always added 'provides' for any package in a release group but I stopped and thought this time, isn't that only used for stuff that might be a dependency? I think I only started doing it as the first versioning example was Quoth. I have not added provides here but let me know if it needs to be